refactor: вынес логику dat.gui
вынес логику dat.gui в отдельный файл
